requirements.txt
altair==5.5.0
attrs==25.3.0
bleach==6.2.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
db-dtypes==1.4.2
fonttools==4.56.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.2
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-cloud-bigquery==3.30.0
google-cloud-core==2.4.3
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.1
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
idna==3.10
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kaggle==1.7.4.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
narwhals==1.30.0
numpy==2.2.3
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
pandas-gbq==0.28.0
pillow==11.1.0
proto-plus==1.26.1
protobuf==5.29.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydata-google-auth==1.9.1
pydeck==0.9.1
PyMySQL==1.1.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.1
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.23.1
rsa==4.9
setuptools==76.0.0
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.39
streamlit==1.43.2
tenacity==9.0.0
text-unidecode==1.3
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
webencodings==0.5.1
